fix(atomix): Unflake multiple ITs
HealthMonitoringTest and DeploymentClusteredTest are both flaky. The
test timesout due to the outerrule (timeout) wrapped around the
embeddedbrokerrule.

At least for the health monitoring test it was clear that a lot of time
was spend waiting for the broker to finish, because the raft partition
server was already stopped inside the test and could then later not be
stopped by the after of the embeddedbrokerrule.

This commit allows calling the shutdown method of DefaultRaftServer
multiple times by keeping track of a stopped flag. The server is only
allowed to stop if it is started and not already stopped. If it is
already stopped it can just return.

Co-authored-by: Peter Ihme <peter.ihme@camunda.com>